move interrupt functions from core_hal.h to interrupts_hal.h
